+++ core/src/java/com/pb/mtctm2/abm/ctramp/Tour.java
@@ -394,7 +394,7 @@ public class Tour
 
     public String getTourWindow( String purposeAbbreviation ) {
         String returnString = String.format("      %5s:     |", purposeAbbreviation );
-        int[] windows = perObj.getTimeWindows();
+        short[] windows = perObj.getTimeWindows();
         for ( int i=1; i < windows.length; i++ ) {
             String tempString = String.format("%s", i >= tourDepartPeriod && i <= tourArrivePeriod ? purposeAbbreviation : "    " );
             if ( tempString.length() == 2 || tempString.length() == 3 )
